+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/abstracts/MediaGroupContent.kt
@@ -1,7 +1,8 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.MediaGroupMemberInputMedia
+import com.github.insanusmokrassar.TelegramBotAPI.types.files.abstracts.TelegramMediaFile
 
-interface MediaGroupContent : MessageContent {
+interface MediaGroupContent<T : TelegramMediaFile> : MediaContent<T>, MessageContent {
     fun toMediaGroupMemberInputMedia(): MediaGroupMemberInputMedia
 }